+++ packages/core/src/Reporter.ts
@@ -160,6 +160,13 @@ export default class Reporter<Options extends object = {}> extends Plugin<Option
   };
 
   /**
+   * Return true if the user's terminal supports color.
+   */
+  hasColorSupport(level: number = 1): boolean {
+    return chalk.supportsColor && chalk.supportsColor.level >= level;
+  }
+
+  /**
    * Create an indentation based on the defined length.
    */
   indent(length: number = 0): string {
@@ -202,6 +209,10 @@ export default class Reporter<Options extends object = {}> extends Plugin<Option
     type: ColorType = 'default',
     modifiers: ('reset' | 'bold' | 'dim' | 'italic' | 'underline' | 'inverse' | 'hidden')[] = [],
   ): string {
+    if (!this.hasColorSupport()) {
+      return message;
+    }
+
     const color = this.getColorPalette()[type];
     let out = color.charAt(0) === '#' ? chalk.hex(color) : chalk[color as Color];