+++ packages/reporter-nyan/src/index.ts
@@ -1,0 +1,6 @@
+/**
+ * @copyright   2017-2019, Miles Johnson
+ * @license     https://opensource.org/licenses/MIT
+ */
+
+export { default } from './NyanReporter';