+++ packages/core/CHANGELOG.md
@@ -6,6 +6,7 @@
 - Added `Tool#isPluginEnabled` method to check if a plugin by type has been enabled.
 - Added `Plugin#blueprint`, `Reporter#blueprint`, and `Routine#blueprint` as a means to validate and
   build options passed to the constructor.
+- Added `Reporter#hasColorSupport` to check for terminal color support.
 - Registered plugins can now define a list of custom NPM `scopes` to lookup modules in.
 
 #### 🛠 Internal